Move `FileSlice` and `SourceFileMetadata` to `semantic_layer_components` in artifacts
